+++ TM4C123_ADC/xDriver_MCU/GPIO/Driver/Intrinsics/Interrupt/InterruptRoutine/xSource/GPIO_InterruptRoutine_Vector_GPIOE.c
@@ -47,7 +47,7 @@ void GPIOE__vIRQVectorHandler(void)
          }
      }
 
-     if(GPIO_enBUS_APB ==u32RegBUS)
+     if((uint32_t)GPIO_enBUS_APB ==u32RegBUS)
      {
          u32Reg=GPIOE_APB_GPIOMIS_R;
          if(u32Reg & GPIO_enPIN_0)
