+++ TM4C123_ADC/xDriver_MCU/TIMER/Driver/Intrinsics/Primitives/xSource/TIMER_Ready.c
@@ -26,7 +26,7 @@
 #include <xDriver_MCU/TIMER/Driver/Intrinsics/Primitives/xHeader/TIMER_Ready.h>
 #include <xDriver_MCU/TIMER/Peripheral/TIMER_Peripheral.h>
 
-static SYSCTL_nPERIPHERAL SYSCTL_VECTOR_TIMER[(uint32_t)(TIMER_enLETTER_MAX+1u)][(uint32_t)(TIMER_enMODULE_NUM_MAX+1u)]={
+static SYSCTL_nPERIPHERAL SYSCTL_VECTOR_TIMER[(uint32_t)TIMER_enLETTER_MAX+1u][(uint32_t)TIMER_enMODULE_NUM_MAX+1u]={
                                                {SYSCTL_enTIMER0,SYSCTL_enTIMER1,SYSCTL_enTIMER2,SYSCTL_enTIMER3,SYSCTL_enTIMER4,SYSCTL_enTIMER5},
                                                {SYSCTL_enWTIMER0,SYSCTL_enWTIMER1,SYSCTL_enWTIMER2,SYSCTL_enWTIMER3,SYSCTL_enWTIMER4,SYSCTL_enWTIMER5}
                                             };
@@ -35,9 +35,9 @@ void TIMER__vSetReady(TIMER_nMODULE enModule)
     SYSCTL_nPERIPHERAL enPeripheral=SYSCTL_enTIMER0;
     uint32_t u32Number= (uint32_t) enModule & 0x7u;
      uint32_t u32Wide= ((uint32_t) enModule>>16u) & 0x1u;
-    if(TIMER_enMISC_MAX<u32Number)
+    if((uint32_t)TIMER_enMISC_MAX<u32Number)
     {
-        u32Number=TIMER_enMISC_MAX;
+        u32Number=(uint32_t)TIMER_enMISC_MAX;
     }
     enPeripheral=SYSCTL_VECTOR_TIMER[u32Wide][u32Number];
     SYSCTL__vSetReady(enPeripheral);
@@ -48,9 +48,9 @@ void TIMER__vClearReady(TIMER_nMODULE enModule)
     SYSCTL_nPERIPHERAL enPeripheral=SYSCTL_enTIMER0;
     uint32_t u32Number= (uint32_t) enModule & 0x7u;
      uint32_t u32Wide= ((uint32_t) enModule>>16u) & 0x1u;
-    if(TIMER_enMISC_MAX<u32Number)
+    if((uint32_t)TIMER_enMISC_MAX<u32Number)
     {
-        u32Number=TIMER_enMISC_MAX;
+        u32Number=(uint32_t)TIMER_enMISC_MAX;
     }
     enPeripheral=SYSCTL_VECTOR_TIMER[u32Wide][u32Number];
     SYSCTL__vClearReady(enPeripheral);
@@ -62,9 +62,9 @@ TIMER_nREADY TIMER__enIsReady(TIMER_nMODULE enModule)
     SYSCTL_nPERIPHERAL enPeripheral=SYSCTL_enTIMER0;
     uint32_t u32Number= (uint32_t) enModule & 0x7u;
     uint32_t u32Wide= ((uint32_t) enModule>>16u) & 0x1u;
-    if(TIMER_enMISC_MAX<u32Number)
+    if((uint32_t)TIMER_enMISC_MAX<u32Number)
     {
-        u32Number=TIMER_enMISC_MAX;
+        u32Number=(uint32_t)TIMER_enMISC_MAX;
     }
     enPeripheral=SYSCTL_VECTOR_TIMER[u32Wide][u32Number];
     enReady=(TIMER_nREADY)SYSCTL__enIsReady(enPeripheral);
